+++ src/components/select/hooks/useOutsideClickClose.ts
@@ -1,31 +1,33 @@
 import { useEffect } from 'react';
 
 type UseOutsideClickClose = {
-	isMenuOpen: boolean;
-	onChange: (newValue: boolean) => void;
-	onClose?: () => void;
-	rootRef: React.RefObject<HTMLDivElement>;
+  isMenuOpen: boolean;
+  onChange: (newValue: boolean) => void;
+  onClose?: () => void;
+  rootRef: React.RefObject<HTMLDivElement>;
 };
 
 export const useOutsideClickClose = ({
-	isMenuOpen,
-	rootRef,
-	onClose,
-	onChange,
+  isMenuOpen,
+  rootRef,
+  onClose,
+  onChange,
 }: UseOutsideClickClose) => {
-	useEffect(() => {
-		const handleClick = (event: MouseEvent) => {
-			const { target } = event;
-			if (target instanceof Node && !rootRef.current?.contains(target)) {
-				isMenuOpen && onClose?.();
-				onChange?.(false);
-			}
-		};
+  useEffect(() => {
+    if (!isMenuOpen) return; // не навешиваем обработчик, если меню закрыто
 
-		window.addEventListener('mousedown', handleClick);
+    const handleClick = (event: MouseEvent) => {
+      const { target } = event;
+      if (target instanceof Node && !rootRef.current?.contains(target)) {
+        onClose?.();
+        onChange(false);
+      }
+    };
 
-		return () => {
-			window.removeEventListener('mousedown', handleClick);
-		};
-	}, [onClose, onChange, isMenuOpen]);
+    window.addEventListener('mousedown', handleClick);
+
+    return () => {
+      window.removeEventListener('mousedown', handleClick);
+    };
+  }, [isMenuOpen, onClose, onChange, rootRef]);
 };
